soft rotate ftw

+++ mjpg-streamer-experimental/plugins/input_uvc/jpeg_utils.h
@@ -1,2 +1,2 @@
 int compress_image_to_jpeg(struct vdIn *vd, unsigned char *buffer, int size, int quality);
-int rotate_jpeg(unsigned char *buffer, int size);
+unsigned long rotate_jpeg(unsigned char *buffer, int size);
